requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
asgiref==3.5.2
async-timeout==4.0.2
attrs==22.2.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.11.1
certifi==2022.12.7
charset-normalizer==2.1.1
colorama==0.4.6
dj-database-url==1.2.0
Django==4.1.1
django-bootstrap-v5==1.0.11
django-environ==0.9.0
django-utils-six==2.0
frozenlist==1.3.3
gunicorn==20.1.0
idna==3.4
multidict==6.0.4
openai==0.26.4
Pillow==9.3.0
psycopg2-binary==2.9.5
requests==2.28.1
soupsieve==2.3.2.post1
sqlparse==0.4.2
tqdm==4.64.1
tzdata==2022.2
urllib3==1.26.14
whitenoise==6.3.0
yarl==1.8.2
